+++ CHANGELOG.md
@@ -59,6 +59,7 @@
 * React Query setup and apply to Search. Refs [UILD-676].
 * Add request building and response transforming strategies for Search. Update UI to support segment hierarchy. Refs [UILD-683].
 * Add results formatting strategies for Search. Update UI to support search results for different types of search. Refs [UILD-683].
+* Add custom workspace profile settings application. Refs [UILD-684].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -118,6 +119,7 @@
 [UILD-675]:https://folio-org.atlassian.net/browse/UILD-675
 [UILD-676]:https://folio-org.atlassian.net/browse/UILD-676
 [UILD-683]:https://folio-org.atlassian.net/browse/UILD-683
+[UILD-684]:https://folio-org.atlassian.net/browse/UILD-684
 
 ## 1.0.5 (2025-04-30)
 * Fixed incorrect behavior when navigating between duplicated resources. Fixes [UILD-553].